Rewrite plugin to catch keystrokes instead of editor changes
Turned out watching the editor for changes is too error prone so the plugin
is changed to listen to keystrokes instead. Each captured keystroke is
counted as 1 XP. This counts system keystrokes like saving and stuff, but
it should still be more accurate than before.

Also fixed unsubscribing properly from events when the plugin is disabled.

Fixes #3, #4 I hope